chore: headquarter 도메인 application 패키지명 오타 수정
