+++ src/test/java/com/goorm/friendchise/domain/headquarter/application/KakaoApiServiceTest.java
