+++ src/test/java/com/goorm/friendchise/domain/headquarter/application/OpenAiApiServiceTest.java
